bunch of restructuring (again) to prevent circular imports
