+++ rewardgym/stimuli/__init__.py
@@ -1,7 +1,6 @@
-from itertools import permutations
-
 from .alien_images import draw_alien
 from .default_images import (
+    STIMULUS_DEFAULTS,
     generate_stimulus_properties,
     make_card_stimulus,
     mid_stimuli,
@@ -30,25 +29,5 @@ __all__ = [
     "mid_stimuli",
     "make_card_stimulus",
     "generate_stimulus_properties",
+    "STIMULUS_DEFAULTS",
 ]
-
-
-bg_color = (240, 240, 240)
-shapes = ["square", "circle", "triangle_u", "triangle_d", "diamond"]
-shapes_perm = shapes[:] + list(permutations(shapes, r=2))
-patterns = [(2, 3), (4, 6), (1, 2)]
-
-colors = [
-    (1, 25, 89),
-    (16, 63, 96),
-    (28, 90, 98),
-    (60, 109, 86),
-    (104, 123, 62),
-    (157, 137, 43),
-    (210, 147, 67),
-    (248, 161, 123),
-    (253, 183, 188),
-    (250, 204, 250),
-] + [bg_color]  # batlow 10
-
-STIMULUS_DEFAULTS = {"shapes": shapes_perm, "colors": colors, "patterns": patterns}
